# backend/docs/openapi/OverfitterAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "Overfitter",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "localhost:4000"
    },
    {
      "url": "overfitter.jamkelley.com/api"
    }
  ],
  "paths": {
...
    "/outfits": {
      "get": {
        "summary": "Get All Outfits",
        "operationId": "getAllOutfits",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Outfit",
        "operationId": "createOutfit",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Outfits",
        "operationId": "deleteAllOutfits",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/outfits/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Outfit by Id",
        "operationId": "getOutfitById",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Outfit by Id",
        "operationId": "deleteOutfitById",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Outfit by Id",
        "operationId": "updateOutfitById",
        "tags": [
          "Outfits"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/outfits/search": {
      "post": {
        "summary": "Search for Outfit Item",
        "operationId": "searchOutfitItem",
        "tags": [
          "Outfits"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/EntitySearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OutfitResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/tops": {
      "get": {
        "summary": "Get All Top Items",
        "operationId": "getAllTopItems",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "post": {
        "summary": "Create Top Item",
        "operationId": "createTopItem",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete All Top Items",
        "operationId": "deleteAllTopItems",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/tops/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/idParam"
        }
      ],
      "get": {
        "summary": "Get Top Item by Id",
        "operationId": "getTopItemById",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "delete": {
        "summary": "Delete Top Item by Id",
        "operationId": "deleteTopById",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      },
      "put": {
        "summary": "Update Top Item by Id",
        "operationId": "updateTopItemById",
        "tags": [
          "Tops"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
    "/tops/search": {
      "post": {
        "summary": "Search for Top Item",
        "operationId": "searchTopItem",
        "tags": [
          "Tops"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ClothingItemSearchBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TopResponseList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "5XX": {
            "$ref": "#/components/responses/UnexpctedError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "OutfitResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Outfit"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "OutfitResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Outfit"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "TopResponse": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/WithId"
                        },
                        {
                          "$ref": "#/components/schemas/Top"
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "TopResponseList": {
        "description": "Ok",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DatalessResponse"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/WithId"
                          },
                          {
                            "$ref": "#/components/schemas/Top"
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      },
      "UnexpctedError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorType"
            }
          }
        }
      }
    },
    "schemas": {
      "WithId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
      "DatalessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "item_status": {
        "type": "string",
        "enum": [
          "CLEAN",
          "DIRTY",
          "WORN",
          "OK",
          "UNKNOWN"
        ]
      },
      "item_condition": {
        "type": "string",
        "enum": [
          "NEW",
          "LIKE_NEW",
          "GOOD",
          "FAIR",
          "POOR",
          "TRASH",
          "UNKNOWN"
        ]
      },
      "size_us_letter": {
        "type": "string",
        "enum": [
          "5XS",
          "4XS",
          "3XS",
          "2XS",
          "XS",
          "S",
          "M",
          "L",
          "XL",
          "2XL",
          "3XL",
          "4XL",
          "5XL",
          "UNKNOWN"
        ]
      },
      "Entity": {
        "type": "object",
        "properties": {
          "uriImage": {
            "type": "string",
            "pattern": "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)",
            "example": "https://www.pngmart.com/files/11/Rickrolling-PNG-Transparent-Image.png"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "timestampAddedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "timestampLastModifiedRFC": {
            "type": "string",
            "format": "date-time"
          },
          "rating": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          }
        }
      },
      "ClothingItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entity"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "sizeUSLetter": {
                "$ref": "#/components/schemas/size_us_letter"
              },
              "sizeUSNumber": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "brand": {
                "type": "string"
              },
              "timestampPurchasedRFC": {
                "type": "string",
                "format": "date-time"
              },
              "itemCondition": {
                "$ref": "#/components/schemas/item_condition"
              },
              "itemStatus": {
                "$ref": "#/components/schemas/item_status"
              },
              "numberOfWears": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "wearsLeftBeforeDirty": {
                "type": "integer",
                "minimum": 0,
                "maximum": 1000000
              },
              "primaryColor": {
                "type": "string"
              },
              "secondaryColor": {
                "type": "string"
              },
              "accentColor": {
                "type": "string"
              },
              "pattern": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "outfit_type": {
        "type": "string",
        "enum": [
          "FORMAL",
          "CASUAL",
          "BUSINESS",
          "BUSINESS_CASUAL",
          "SPORT",
          "SLEEP",
          "UNKNOWN"
        ]
      },
      "Outfit": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entity"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/outfit_type"
              },
              "topId": {
                "type": "integer",
                "minimum": 0
              },
              "bottomId": {
                "type": "integer",
                "minimum": 0
              },
              "headId": {
                "type": "integer",
                "minimum": 0
              },
              "feetId": {
                "type": "integer",
                "minimum": 0
              },
              "wristId": {
                "type": "integer",
                "minimum": 0
              },
              "rating": {
                "type": "integer",
                "minimum": 0,
                "maximum": 5
              }
            }
          }
        ]
      },
      "top_type": {
        "type": "string",
        "enum": [
          "TEE",
          "SWEAT",
          "DRESS",
          "FLANNEL",
          "UNDER",
          "SLEEP",
          "CASUAL",
          "WORKOUT",
          "UNKNOWN"
        ]
      },
      "Top": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ClothingItem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "$ref": "#/components/schemas/top_type"
              },
              "sizeChestInchesLow": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "sizeChestInchesHigh": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "sizeNeckInchesLow": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              },
              "sizeNeckInchesHigh": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              },
              "sizeWaistInchesLow": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "sizeWaistInchesHigh": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "sizeSleeveInches": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              }
            }
          }
        ]
      },
...
      "ErrorType": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "errorMessage": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "nullable": true
          }
        }
      }
    },
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "description": "ID of the Item",
        "required": true,
        "schema": {
          "type": "integer"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
...
      "EntitySearchBody": {
        "description": "A JSON object containing search info for an Entity",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "timestampAddedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampAddedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedBeginRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "timestampModifiedEndRFC": {
                  "type": "string",
                  "format": "date-time"
                },
                "rating": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "ClothingItemSearchBody": {
        "description": "A JSON object containing search info for a ClothingItem",
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/requestBodies/EntitySearchBody"
                },
                {
                  "type": "object",
                  "properties": {
                    "sizeUSLetter": {
                      "type": "string"
                    },
                    "sizeUSNumber": {
                      "type": "integer"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "timestampPurchasedBeginRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "timestampPurchasedEndRFC": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "itemCondition": {
                      "type": "string"
                    },
                    "itemStatus": {
                      "type": "string"
                    },
                    "numberOfWearsMin": {
                      "type": "integer"
                    },
                    "numberOfWearsMax": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMin": {
                      "type": "integer"
                    },
                    "wearsLeftBeforeDirtyMax": {
                      "type": "integer"
                    },
                    "primaryColor": {
                      "type": "string"
                    },
                    "secondaryColor": {
                      "type": "string"
                    },
                    "accentColor": {
                      "type": "string"
                    },
                    "pattern": {
                      "type": "string"
                    },
                    "itemType": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}